Exchange dependency RPy for fisher.
RPy, Python bindings to R, is a problematic dependency to satisfy due to
requiring installation of R and its source files. Additionally, the RPy
package that ships with Ubuntu 10.10 is broken. The use of RPy simply
for access to an implementation of Fisher's Exact Test was very heavy.

The fisher Python package, available from PyPI at
http://pypi.python.org/pypi/fisher, is easily installed through pip. It
is fast, being written in Cython, and beats using Python bindings to R.
Comparison showed its calculated values nearly identical to R's
fisher.test(). Thus, it is suitable as a replacement and a more
palatable dependency, overall.

I have further updated the installation documentation in README.rst,
including listing the new dependency, and describing installation
through pip and using virtualenv.

git-svn-id: svn+ssh://cowcreamer.cs.vt.edu/home/murali/svnrepo/src/python/BiologicalProcessNetworks/trunk@3579 1fbec0df-f6b4-4bca-857c-de90862b47bd