Add README.html, processed from README.rst.
Not sure I want to keep this file around, or in the top directory,
however, I'm the only person in our group that has docutils installed,
and the generated HTML is easier to read than the ReST file. Just
remember that this file will fall out of sync if someone makes changes
to README.rst and does not run rst2html on that file to re-generate the
HTML version.

git-svn-id: svn+ssh://cowcreamer.cs.vt.edu/home/murali/svnrepo/src/python/BiologicalProcessNetworks/trunk@3601 1fbec0df-f6b4-4bca-857c-de90862b47bd