Add data classes for term-based model.
``Symmetrical2dArray`` is a supporting class that subclasses numpy's
array class and ensures that for array ``a``, ``a[i, j] == a[j, i]``.

``symzeros`` is a convenience function similar to ``numpy.zeros`` but
returning a ``Symmetrical2dArray`` instead.

``AnnotatedInteractions2dArray`` is a new structure akin to
``AnnotatedInteractionsGraph``, but with the two-dimensional symmetrical
array at its core instead of a dictionary.

git-svn-id: svn+ssh://cowcreamer.cs.vt.edu/home/murali/svnrepo/src/python/BiologicalProcessNetworks/trunk@3717 1fbec0df-f6b4-4bca-857c-de90862b47bd