Adds CLI options, logic for terms-based model.
The new option is "--terms-based", which causes the underlying Markov
chain to be ``TermsBasedMarkovChain``.

The ``main()`` function now also calculates component sizes for the
potential process network from all co-annotating pairs, which allows the
user to check to see if their input is suitable (e.g., for the
terms-based approach, which currently could break reachability in the
Markov chain). This is done through the ``check_link_components()``
helper function. Maybe this should have been put in ``structures.py``,
but right now it works.

git-svn-id: svn+ssh://cowcreamer.cs.vt.edu/home/murali/svnrepo/src/python/BiologicalProcessNetworks/trunk@3724 1fbec0df-f6b4-4bca-857c-de90862b47bd