Adds model that includes intra-term interactions.
The model builds upon the terms-based model. The intra-term interactions
are those for which the incident genes both share some annotation term.
If that term participates in one or more links marked as selected, then
those intra-term interactions will also be marked as selected if they
have not yet already been.

Adds two new classes to enact the new model,
``IntratermInteractions2dArray``, a data structure mapping terms and
links to their interactions, and ``IntraTermsAndLinksState``, which is
similar to ``TermsAndLinksState``, but with the shift in calculation of
selected interactions to include intra-term interactions.

To use the new model, invoke the new "--intraterms" option along with
the "--terms-based" option.  (NOTE: At the moment there's no warning to
the user if they do not supply "--terms-based" along with
"--intraterms", although it is noted in the help-string documentation.)

git-svn-id: svn+ssh://cowcreamer.cs.vt.edu/home/murali/svnrepo/src/python/BiologicalProcessNetworks/trunk@3738 1fbec0df-f6b4-4bca-857c-de90862b47bd