Allows user to specify seed links.
The user can specify a two-column CSV file containing the links he or
she would like to use as a seed for the Markov chain by providing the
new "--seed-links" option.

I also renamed "selected_links" ("selected_links_indices") to
"seed_links" ("seed_links_indices") in the initialization methods of
various states to better represent the intention of the parameter to
users.

git-svn-id: svn+ssh://cowcreamer.cs.vt.edu/home/murali/svnrepo/src/python/BiologicalProcessNetworks/trunk@3778 1fbec0df-f6b4-4bca-857c-de90862b47bd