Updated the simulated annealing algorithm to begin with a much larger temperature; 100,000. Updated the degradation of the temperature. Degrading the temperature still relies on using the number of steps entered by the user, but will not anneal to the specific amount. Rather they are directly correlated. See the source code for a further description. Also, the algorithm now anneals while the temperature is > 0.1 as opposed to > 0 due to the new degradation of the temperature.
The likelihood of accepting a worse state now correctly relies on delta E and the temperature, as opposed to only the temperature. However, due to the bug of not creating a new state, delta E is always 0 and the worse state is always accepted. There is also a math range overflow bug that needs be addressed when computing the likelihood of accepting a worse state based on the temperature and delta E.

git-svn-id: svn+ssh://cowcreamer.cs.vt.edu/home/murali/svnrepo/src/python/BiologicalProcessNetworks/trunk@3796 1fbec0df-f6b4-4bca-857c-de90862b47bd