Adds link state where terms independently selected.
Adds ``SelectableTermsAndLinksState``, a class in which terms may be
selected indpendent of links (i.e., some terms may be selected but have
no links to any other selected terms).

Refactors a number of methods in ``TermsAndLinksState`` to facilitate
the production of the new grandchild subclass. In particular, it now
tracks the number of links in which each term participates
(``_term_links_counts``), and tracks this appropriately through the new
methods ``_select_terms_via_link()`` and ``_unselect_terms_via_link()``.

It also renames ``_term_selections`` to ``term_selections``, as this
should be an attribute inspected by a new recorder class, which will
tally for each iteration in the Markov chain which terms were selected.

Simplifies some of ``IntraTermsAndLinksState`` methods.

git-svn-id: svn+ssh://cowcreamer.cs.vt.edu/home/murali/svnrepo/src/python/BiologicalProcessNetworks/trunk@3799 1fbec0df-f6b4-4bca-857c-de90862b47bd