Rewrite of terms-based state recorder.
Renames ``TermsStateRecorder`` to ``TermsBasedStateRecorder``.
Eliminates ``DetailedTermsStateRecorder`` by combining it into
``TermsBasedStateRecorder``. Since the user can choose to have all the
transitions data output to a bzip2 compressed file, and it's always been
helpful to have the full information about each state in the transition
data, we might as well just use it as the standard.

This state recorder now records the selection states of the terms. It
also provides a new method, ``write_terms_probabilities``, to output the
probability of each term at the end of a run. It is analagous to
``write_links_probabilities``.

git-svn-id: svn+ssh://cowcreamer.cs.vt.edu/home/murali/svnrepo/src/python/BiologicalProcessNetworks/trunk@3818 1fbec0df-f6b4-4bca-857c-de90862b47bd