Adds fixed distributions parameters state.
The user can invoke this new parameters state in the model by issuing
the "--fixed-distributions" option on the command line.

This parameters state does not derive its distribution from the number
of terms and potential links, but instead is fixed to be the range
[0.025, 0.5], at increments of 0.025. The result is that the Markov
chain can reach much higher values for the term and link prior than
previously available. Our reasoning is that these values were far too
small before.

``FixedDistributionParametersState`` is the new class providing this
parameter state.

git-svn-id: svn+ssh://cowcreamer.cs.vt.edu/home/murali/svnrepo/src/python/BiologicalProcessNetworks/trunk@3825 1fbec0df-f6b4-4bca-857c-de90862b47bd