Adds more comments on state transition logic.
PJW had some questions about the calculation which I had difficulty
explaining earlier, so I sat down and wrote detailed comments on the
calculation of the transition ratio, and how this transition ratio is
used to determine whether to accept the proposed state or not.

Additionally, I made the code more explicit to indicate that we always
accept the proposed state if it is more favorable, and otherwise compare
it to a probability to determine if we will reject or accept the
proposed state.

Comments appear in ``PLNMarkovChain.next_state()``.

git-svn-id: svn+ssh://cowcreamer.cs.vt.edu/home/murali/svnrepo/src/python/BiologicalProcessNetworks/trunk@3863 1fbec0df-f6b4-4bca-857c-de90862b47bd