Removes starting with random sample of links.
If the user provides no seed links, the starting state for the links
state will be that no links (and terms) will be selected. This is in
contrast to before, where the overall state would choose a random sample
of links to select at the start if no seed links were provided. My hope
is this significantly improves the performance of the chain, as the
random sampling was leading to too terrible a starting state.

git-svn-id: svn+ssh://cowcreamer.cs.vt.edu/home/murali/svnrepo/src/python/BiologicalProcessNetworks/trunk@3864 1fbec0df-f6b4-4bca-857c-de90862b47bd