Adds gene retrieval methods to AnnotatedInteractions2dArray.
In preparation for a gene-based model, this adds several methods for
dealing with genes annotated by terms. The most important method is
``AnnotatedInteractions2dArray.get_annotated_genes()``, which returns a
list of indices to genes which a term annotates. These indices should
then be used to access an array in the links state class to toggle the
selection state of those genes.

git-svn-id: svn+ssh://cowcreamer.cs.vt.edu/home/murali/svnrepo/src/python/BiologicalProcessNetworks/trunk@3894 1fbec0df-f6b4-4bca-857c-de90862b47bd