Adds genes-based overlap model.
This new model measures overlap of terms based on gene content rather
than intra-term interactions. It is essentially an encapsulation of
Bauer et al.'s 2010 "GOing Bayesian" model into the search for process
links.

The new model is accessed with the command line option '--genes-based',
which must be issued with '--terms-based'.

New classes are:

* ``TermsParametersState``, which is not to be confused with, but
  probably will be confuled with, ``TermPriorParametersState``.
  ``TermsParametersState`` includes false-positive and false-negative
  rates specifically for terms selecting genes à la the Bauer paper, and
  the logical equivalents of the rates for links.
* ``GenesBasedTermsAndLinksState``, which tracks the overlap of terms by
  genes rather than intra-term interations.
* ``GenesBasedOverallState``, which calculates the probability of the
  links state now by including the probability of the terms based on
  their overlap at a genes level.
* ``GenesBasedMarkovChain``, the Markov chain which uses the new
  ``GenesBasedOverallState``.

I had to rename ``IndependentTermsAndLinksState`` to
``IndependentIntraTermsAndLinksState`` and create a different
``IndependentTermsAndLinksState`` so that
``GenesBasedTermsAndLinksState`` could inherit only the methods for link
selection and unselection that didn't attempt to record intraterm
interactions. Hopefully ``IndependentIntraTermsAndLinksState`` still
works as ``IndependentTermsAndLinksState`` did before.

git-svn-id: svn+ssh://cowcreamer.cs.vt.edu/home/murali/svnrepo/src/python/BiologicalProcessNetworks/trunk@3895 1fbec0df-f6b4-4bca-857c-de90862b47bd