Interactions as indices, contextual interaction counts.
Interactions were being tracked as actual pairs of genes, which was
causing some overhead in tracking the number of times each interaction
was selected. The data structure doing that previously was a dictionary
that grew large and was fairly expensive to copy (where copying was done
each time a new link transition was proposed).

By representing interactions as integer indices, the code tracks their
selection status through arrays which require less memory and copy more
quickly. This will help with larger interaction data sets such as the
human MiMI network.

I have also added more contextual accounting for the number of
interactions. For models that are based on co-annotated interactions,
only the co-annotated interactions are included in the count of number
of interactions and active interactions. For models that also include
intra-term interactions, these interactions are also included in the
counts (which should give back every interaction, since each gene has at
least one annotation and is incident on at least one interaction). The
idea here is that some interactions could never be covered by the BPN
because they lacked co-annotations (solely intra-term interactions).

git-svn-id: svn+ssh://cowcreamer.cs.vt.edu/home/murali/svnrepo/src/python/BiologicalProcessNetworks/trunk@3996 1fbec0df-f6b4-4bca-857c-de90862b47bd