Bugfix for redundant interactions in co-annotations.
I neglected that the same set of pairs of annotation can come up twice
when doing the cross-product of annotations for the two incident genes.
As a result of storing the mapping of co-annotations to lists of
interactions instead of sets of them, some interactions appeared twice,
which would affect selection and unselection tallies for the individual
interactions, although not the overall tallies.

git-svn-id: svn+ssh://cowcreamer.cs.vt.edu/home/murali/svnrepo/src/python/BiologicalProcessNetworks/trunk@3997 1fbec0df-f6b4-4bca-857c-de90862b47bd