Correctly report log-likelihood of overall state.
Earlier, I was incorporating the number of neighboring states into the
reported log-likelihood of an overall state in the Markov chain. This
removes the neighboring states from the log-likelihood calculation and
correctly places it in the calculation of the transition ratio.

Adds a new method ``PLNMarkovChain.calc_log_transition_ratio()``.

git-svn-id: svn+ssh://cowcreamer.cs.vt.edu/home/murali/svnrepo/src/python/BiologicalProcessNetworks/trunk@4156 1fbec0df-f6b4-4bca-857c-de90862b47bd