Fixes miscalculation of active interactions bug.
I had chosen the wrong way to include interactions in the
``AnnotatedInteractionsGraph._coannotated_and_intraterm_indices`` list,
using a logical AND instead of a logical OR to assign indices to that
list.

Additionally, I had missed changes to call the appropriate methods for
getting the number of interactions and getting the active interactions
from the annotated interactions data structures.

git-svn-id: svn+ssh://cowcreamer.cs.vt.edu/home/murali/svnrepo/src/python/BiologicalProcessNetworks/trunk@4157 1fbec0df-f6b4-4bca-857c-de90862b47bd