Refactors recorders to consume less memory.
mcmcbpn was consuming a tremendous amount of memory because it was
storing a record for every state in system memory, only writing the data
to disk at the end of the run. This changeset refactors the recorder
classes to write the transitions to disk and flush the records after
recording data for a sufficient number of states. mcmcbpn should now run
very reasonably on a fairly modern machine.

Further, I have decoupled the recorder classes from the Markov chain
classes by removing the need to know the parameters and all their
possible values prior to recording states. (This had caused the need to
instantiate the recorders within the Markov chain instance.)

git-svn-id: svn+ssh://cowcreamer.cs.vt.edu/home/murali/svnrepo/src/python/BiologicalProcessNetworks/trunk@4165 1fbec0df-f6b4-4bca-857c-de90862b47bd