Option --detailed-transitions now heeded.
mcmcbpn was always outputting recording and outputting detailed
information about each state in the steps, save for the simplest model.
This fixes this, so now the transitions is restricted to just the
transition type, log transition ratio, log state likelihood, whether the
transition was accepted, and what the log of the rejection probability
was.

If the user provides the ``--detailed-transitions`` flag, the program
records the more verbose details about each state, such as how many
links are in the BPN, the number of active and inactive interactions
covered by the links, etc.

It was important to get this option working, because for runs of 10M
steps, even the bzip2-compressed transitions files were nearly 0.2 GB.

git-svn-id: svn+ssh://cowcreamer.cs.vt.edu/home/murali/svnrepo/src/python/BiologicalProcessNetworks/trunk@4166 1fbec0df-f6b4-4bca-857c-de90862b47bd