Adds "--relaxed-coannotations".
When the user provides this flag, the program uses a more relaxed
definition of "co-annotation", which is the original definition: one
gene must be annotated with at least one of the two terms, and the other
gene must be annotated with at least the other term. Without this
option, the program uses a more stringent definition of co-annotation
(the re-definition) where both genes may not be annotated with both
terms.

git-svn-id: svn+ssh://cowcreamer.cs.vt.edu/home/murali/svnrepo/src/python/BiologicalProcessNetworks/trunk@4168 1fbec0df-f6b4-4bca-857c-de90862b47bd