Records frequency state information. Two flags have been added to the mcmcbpn, --frequency and --frequency-outfile. The --frequency flag is used by mcmcbpn to record state information and --frequency-outfile is used to specify the file to which frequency information should be written too, default file name being state_frequency.tsv. To record frequency state information, a new state recorder class, FrequencyDetailedArrayStateRecorder , was added to records.py.
Frequency information is output as a tab delimited file as follows. The first column specifies the log likelihood, the second column specifies every visitation count, the third column specifies the partial visitation count, and the last column specifies the probability found by the markov chain of visiting the given state. Every visitation count corresponds to the markov chain recording every state visit regardless of the previous state. Partial visitation count corresponds to the markov chain recording new state visits. That is, if the markov chain stays at the current state, then the partial visitation count for that state will not increase. 

TermsBasedFrequencyStateRecorder is also added at the end of records.py, but since we will not be using the terms based model, it is commented out for now.


git-svn-id: svn+ssh://cowcreamer.cs.vt.edu/home/murali/svnrepo/src/python/BiologicalProcessNetworks/trunk@4188 1fbec0df-f6b4-4bca-857c-de90862b47bd