Use strings to as keys for frequency tallies.
Previously, we were converting all the numpy boolean arrays which
represent the link selections into tuples, and using these tuples as
keys to the frequency tallies dictionaries. These were consuming a
tremendous amount of memory. They were also causing a lot of processing
time in the conversion to a tuple.

Now the numpy arrays are converted into strings using the builtin
``.tostring()`` method for numpy arrays, then later, when we need to
store the likelihood, we re-constitute the arrays using the
``numpy.fromstring()`` function. This appears to reduce memory
consumption to at least one-third that of tuples-based serialization,
and also speeds up the runtime and output.

If more reduction in memory consumption when doing the frequencies
tallies is required in the future, there are some good solutions in this
Stack Overflow question:

http://stackoverflow.com/questions/6694835/efficient-serialization-of-numpy-boolean-arrays

git-svn-id: svn+ssh://cowcreamer.cs.vt.edu/home/murali/svnrepo/src/python/BiologicalProcessNetworks/trunk@4199 1fbec0df-f6b4-4bca-857c-de90862b47bd