Track state frequencies by log-likelihood.
Rather than tracking the frequency of each state by the links selected,
this causes tracking to be simply by the log-likelihood. Since we're
going to bin the data anyway, this seemed appropriate. It's also more
efficient in terms of tracking the states and space in the dictionary.
In particular, it saves a lot of time on computation, since these
likelihoods had to be recomputed previously (although we could have and
should have cached the likelihoods too).

There's also some re-formatting and tidying up of the code of the
in ``FrequencyDetailedArrayStateRecorder``, some aesthetic, but some
because there were more Pythonic ways of doing the same thing.

git-svn-id: svn+ssh://cowcreamer.cs.vt.edu/home/murali/svnrepo/src/python/BiologicalProcessNetworks/trunk@4201 1fbec0df-f6b4-4bca-857c-de90862b47bd