Add serialization of states.
Adds ``ArrayLinksState.serialize_state()``, which serializes the state's
``link_selections`` array as a bit array. This bit array can be used as
a key in the frequency tallies dictionaries of the frequency recorders.
It's quite space-efficient and doesn't seem to cause much slowdown in
the runtime of the program.

Adds ``PLNOverallState.serialize_state()``, which at the moment, only
returns the serialization of the links state. Later we can include some
serialization that incorporates the parameters state, too.

Adds new class ``BoolBitArray`` to ``structures.py``. This makes the bit
arrays hashable structures suitable to use as keys in a dictionary.
(NumPy arrays are not hashable objects, themselves, and so may not be
used as keys in a dictionary.) This code is adapted from Stack Overflow.
See
http://stackoverflow.com/questions/6694835/efficient-serialization-of-numpy-boolean-arrays/6695272#6695272

Adds the external Python package bitarray as a dependency.

git-svn-id: svn+ssh://cowcreamer.cs.vt.edu/home/murali/svnrepo/src/python/BiologicalProcessNetworks/trunk@4205 1fbec0df-f6b4-4bca-857c-de90862b47bd